macos: move old toolbar into ventura file
